Move more generic test functions to test mod
